feat(seeders): add CouponSeeder, DatabaseSeeder, SettingTableSeeder, and UsersTableSeeder to seed the database with initial data
The CouponSeeder is added to seed the 'coupons' table with two initial coupon records. The records include the code, type, value, and status of each coupon.

The DatabaseSeeder is added to seed the application's database. It calls the UsersTableSeeder, SettingTableSeeder, and CouponSeeder to populate the respective tables with initial data.

The SettingTableSeeder is added to seed the 'settings' table with initial settings data. The data includes the description, short description, photo, logo, address, email, and phone number.

The UsersTableSeeder is added to seed the 'users' table with two initial user records. The records include the name, email, password, role, and status of each user.